Added decoding pipeline file
The evaluate_decoder method should be the only public facing method as the rest should only need to be called by evaluate_classical_decoder.

I have created 2 versions of evaluate_classical_decoder. One of them should work, and it takes in a decoder object. The type of the object detemines which decoder is called later on in the function. The decoding is done with a Pauli frame simulation based upon code written by Anthony for evaluating the belief propagation decoder.

I have tried to test my methods, but I've been having a problem with running standard Stabilizer type methods on Stabilizer objects that I created from a Hx and Hz matrix. Pointers on how to fix this would be appreciated!

I have included a CSS class below which includes some methods I use in the decoder.